Feat : Ajout de la fonctionnalité de blocage des action selon les droits

+++ restaurant/views.py
@@ -1,3 +1,4 @@
+from django.contrib.auth.middleware import AuthenticationMiddleware
 from django.http import HttpResponse, Http404
 from django.shortcuts import redirect, get_object_or_404
 from django.template import loader
@@ -15,16 +16,20 @@ def index(request):
 
 
 def create(request):
-    if request.method == 'POST':
-        form = RestaurantForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return redirect('index')
+    user = request.user.is_authenticated
+    if user:
+        if request.method == 'POST':
+            form = RestaurantForm(request.POST)
+            if form.is_valid():
+                form.save()
+                return redirect('index')
+        else:
+            template = loader.get_template("restaurants/add.html")
+            form = RestaurantForm()
+            context = {"form": form}
+            return HttpResponse(template.render(context, request))
     else:
-        template = loader.get_template("restaurants/add.html")
-        form = RestaurantForm()
-        context = {"form": form}
-        return HttpResponse(template.render(context, request))
+        raise Http404
 
 
 def detail(request, pk):
@@ -39,20 +44,28 @@ def detail(request, pk):
 
 
 def update(request, pk):
-    restaurant = Restaurant.objects.get(id=pk)
-    if request.method == 'POST':
-        form = RestaurantForm(request.POST, instance=restaurant)
-        if form.is_valid():
-            form.save()
-            return redirect('index')
+    user = request.user.is_authenticated
+    if user:
+        restaurant = Restaurant.objects.get(id=pk)
+        if request.method == 'POST':
+            form = RestaurantForm(request.POST, instance=restaurant)
+            if form.is_valid():
+                form.save()
+                return redirect('index')
+        else:
+            template = loader.get_template("restaurants/update.html")
+            form = RestaurantForm(instance=restaurant)
+            context = {"form": form, "restaurant": restaurant}
+            return HttpResponse(template.render(context, request))
     else:
-        template = loader.get_template("restaurants/update.html")
-        form = RestaurantForm(instance=restaurant)
-        context = {"form": form, "restaurant": restaurant}
-        return HttpResponse(template.render(context, request))
+        raise Http404
 
 
 def delete(request, pk):
-    restaurant_to_delete = get_object_or_404(Restaurant, id=pk)
-    restaurant_to_delete.delete()
-    return redirect('index')
+    user = request.user.is_authenticated
+    if user:
+        restaurant_to_delete = get_object_or_404(Restaurant, id=pk)
+        restaurant_to_delete.delete()
+        return redirect('index')
+    else:
+        raise Http404
